requirements.txt
torch
sentence-transformers
faiss-gpu
numpy
fastapi
uvicorn
